PR feedback: extract empty table message component
